Moved react form to the components directory as it makes more sense from here on out.
